reduce memory usage

+++ brewlib/Util/ByteCounterStream.cs
@@ -15,7 +15,7 @@ public class ByteCounterStream : Stream
     public override void Flush() { }
 
     public override int Read(byte[] buffer, int offset, int count) => int.MinValue;
-    public override long Seek(long offset, SeekOrigin origin) => int.MinValue;
+    public override long Seek(long offset, SeekOrigin origin) => long.MinValue;
 
     public override void SetLength(long value) => length = value;
     public override void Write(byte[] buffer, int offset, int count) => length += count;
